AI-NLP/requirements.txt
# Core
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LangChain
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0

# PDF Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# NLP
spacy>=3.7.0

# Embeddings & Vector Store
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.0
